refactoring: simplify fixture mechanism for integration_tests
