--- integration-testing/test/star_connected/test_casper_propose_and_deploy.py
@@ -1,6 +1,0 @@
-from tools.profiling import profile
-from tests.casper_propose_and_deploy import casper_propose_and_deploy
-
-@profile
-def test_casper_propose_and_deploy(config, converged_star_network):
-    casper_propose_and_deploy(config, converged_star_network)
