--- integration-testing/test/star_connected/test_convergence.py
@@ -1,7 +1,0 @@
-import logging
-from tools.profiling import profile
-
-@profile
-def test_network_convergence(converged_star_network):
-    logging.info("Network converged successfully")
-    pass
